requirements.txt
alembic==1.14.1
altgraph==0.17.4
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
Flask-Assets==2.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
is-disposable-email==1.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsonify==0.5
Mako==1.3.8
MarkupSafe==3.0.2
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pytesseract==0.3.13
python-dotenv==1.0.1
setuptools==75.8.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
uWSGI==2.0.28
webassets==2.0
Werkzeug==3.1.3
WTForms==3.2.1
wtforms-validators==1.0.0
